FIX: Ensure shape matches template
The reverse of the TransformFixedParameters should match the template dimensions.

This fell through the cracks because the templates tested (MNI152 variants) all have the same
X and Z dimensions. Only when testing on a template with distinct values for shape did this become evident.

Additionally, this removes any `FIXED_PARAMS` checks, as that is template specific and more useful for debugging the
initial implementation.